+++ custom_components/spotifyplus/__init__.py
@@ -117,6 +117,7 @@ SERVICE_SPOTIFY_GET_COVER_IMAGE_FILE:str = 'get_cover_image_file'
 SERVICE_SPOTIFY_GET_EPISODE:str = 'get_episode'
 SERVICE_SPOTIFY_GET_EPISODE_FAVORITES:str = 'get_episode_favorites'
 SERVICE_SPOTIFY_GET_FEATURED_PLAYLISTS:str = 'get_featured_playlists'
+SERVICE_SPOTIFY_GET_ID_FROM_URI:str = 'get_id_from_uri'
 SERVICE_SPOTIFY_GET_IMAGE_PALETTE_COLORS:str = 'get_image_palette_colors'
 SERVICE_SPOTIFY_GET_IMAGE_VIBRANT_COLORS:str = 'get_image_vibrant_colors'
 SERVICE_SPOTIFY_GET_PLAYER_DEVICES:str = 'get_player_devices'
@@ -468,6 +469,13 @@ SERVICE_SPOTIFY_GET_FEATURED_PLAYLISTS_SCHEMA = vol.Schema(
         vol.Optional("timestamp"): cv.string,
         vol.Optional("limit_total", default=0): vol.All(vol.Range(min=0,max=9999)),
         vol.Optional("sort_result"): cv.boolean,
+    }
+)
+
+SERVICE_SPOTIFY_GET_ID_FROM_URI_SCHEMA = vol.Schema(
+    {
+        vol.Required("entity_id"): cv.entity_id,
+        vol.Optional("uri"): cv.string,
     }
 )
 
@@ -1191,11 +1199,8 @@ SERVICE_VOLUME_SET_STEP:str = 'volume_set_step'
 SERVICE_VOLUME_SET_STEP_SCHEMA = vol.Schema(
     {
         vol.Required("entity_id"): cv.entity_id,
-        vol.Required("level", default=0.10): vol.All(vol.Range(min=0.01,max=1.00)),
-        #vol.Required("level", default=0.10): vol.All(float, vol.Range(min=.01,max=.99)),
-        #vol.Required("level", default=0.10): vol.All(vol.Range(min=0.01,max=0.99)),
-        #vol.Required("level", default=0.10): vol.All(float, vol.Range(min=0.01,max=1.00)),
-        #vol.Required("level", default=0.10): vol.All(vol.Coerce(float), vol.Range(min=0.01,max=1.00)),
+        vol.Optional("level", default=None): vol.Any(None, vol.All(vol.Coerce(float), vol.Range(min=0.01, max=1.00))),
+        vol.Optional("level_percent", default=None): vol.Any(None, vol.All(vol.Coerce(int), vol.Range(min=1, max=100)))
     }
 )
 
@@ -1621,7 +1626,8 @@ async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
                     # test token expiration.
                     _logsi.LogVerbose(STAppMessages.MSG_SERVICE_EXECUTE % (service.service, entity.name))
                     level = service.data.get("level")
-                    await hass.async_add_executor_job(entity.service_volume_set_step, level)
+                    level_percent = service.data.get("level_percent")
+                    await hass.async_add_executor_job(entity.service_volume_set_step, level, level_percent)
 
                 else:
                     
@@ -1911,6 +1917,13 @@ async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
                     sort_result = service.data.get("sort_result")
                     _logsi.LogVerbose(STAppMessages.MSG_SERVICE_EXECUTE % (service.service, entity.name))
                     response = await hass.async_add_executor_job(entity.service_spotify_get_featured_playlists, limit, offset, country, locale, timestamp, limit_total, sort_result)
+
+                elif service.service == SERVICE_SPOTIFY_GET_ID_FROM_URI:
+
+                    # get id from uri.
+                    uri = service.data.get("uri")
+                    _logsi.LogVerbose(STAppMessages.MSG_SERVICE_EXECUTE % (service.service, entity.name))
+                    response = await hass.async_add_executor_job(entity.service_spotify_get_id_from_uri, uri)
 
                 elif service.service == SERVICE_SPOTIFY_GET_IMAGE_PALETTE_COLORS:
 
@@ -2971,6 +2984,15 @@ async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
             SERVICE_SPOTIFY_GET_FEATURED_PLAYLISTS,
             service_handle_spotify_serviceresponse,
             schema=SERVICE_SPOTIFY_GET_FEATURED_PLAYLISTS_SCHEMA,
+            supports_response=SupportsResponse.ONLY,
+        )
+
+        _logsi.LogObject(SILevel.Verbose, STAppMessages.MSG_SERVICE_REQUEST_REGISTER % SERVICE_SPOTIFY_GET_ID_FROM_URI, SERVICE_SPOTIFY_GET_ID_FROM_URI_SCHEMA)
+        hass.services.async_register(
+            DOMAIN,
+            SERVICE_SPOTIFY_GET_ID_FROM_URI,
+            service_handle_spotify_serviceresponse,
+            schema=SERVICE_SPOTIFY_GET_ID_FROM_URI_SCHEMA,
             supports_response=SupportsResponse.ONLY,
         )