+++ CHANGELOG.md
@@ -6,6 +6,14 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.169 ] - 2025/09/30
+
+  * Added service `get_id_from_uri` to get the Id portion (e.g. `6vc9OTcyd3hyzabCmsdnwE`) of a Spotify URI value (e.g. `spotify:track:6vc9OTcyd3hyzabCmsdnwE`).
+  * Updated service `volume_set_step` to add the `level_percent` argument, which can be used to express the step level as a percentage (1 to 100) instead of a float value.
+  * Added state refresh call for all add / remove favorites services.
+  * Added folder `automations` under the base `spotifyplus` folder that will house sample automations`spotifywebapipython` package requirement to `spotifywebapipython>=1.0.248`.
+  * Updated `spotifywebapipython` package requirement to `spotifywebapipython>=1.0.249`.
+
 ###### [ 1.0.168 ] - 2025/09/27
 
   * Updated `spotifywebapipython` package requirement to `spotifywebapipython>=1.0.248`.